pyproject.toml
[tool.poetry]
name = "imagenet_service"
version = "0.1.0"
description = "Small project to test Imagenet model deployment."
authors = ["f-PLT <pelletier.f@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
torchvision = "^0.16.1"
pillow = "^10.1.0"
python-multipart = "^0.0.6"
pytest = "^7.4.3"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
pylint = "^3.0.2"
isort = "^5.12.0"
nox = "^2023.4.22"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint]
disable = "C0114,C0115,C0116"

[tool.black]
line-length = 88
target-version = ["py310","py311"]

[tool.isort]
profile = "black"
